+++ chain/wallet.go
@@ -1,7 +1,7 @@
 package chain
 
 type Wallet struct {
-	Address []byte  `json:"address"`
-	Balance float32 `json:"balance"`
-	Nonce   uint32  `json:"nonce"`
+	Address []byte `json:"address"`
+	Balance int64  `json:"balance"`
+	Nonce   uint32 `json:"nonce"`
 }
